[DEV] Add commands for MPP key mapping
Grab feature:

A grabbed button/axis is no longer used by the mapper
Grabbing some controls will automatically grab others (i.e.
axis_left_lr, virt_left_l and virt_left_r are always grabbed if at
least one of them is grabbed)

- grab -> bitfield of buttons/axes to grab. 0xffffffff to grab all
- grab_state -> result of grab, bitfield of actually grabbed
buttons/axes. Count ones to know the actual number of axes/buttons.
The button_state arg contains the state of the grabbed buttons
(1:pressed)
- grab_button_event -> event on a grabbed button (pressed/released)
- grab_axis_event -> event (value change) on a grabbed axis

Mapping feature:

Actions are split in axis actions (1 axis + 0..N buttons) and
button_actions (1..N buttons) enums. To disable a mapping, send an
invalid one for the action (i.e. 0 buttons for a button_action, or an
invalid axis [-1] for an axis_action)

- map_button_action -> set a button combination for the given action,
on the given product.
- map_axis_action -> set an axis, and an optionnal button combination
for the given action, on the given product.
- button_mapping_item -> item of the button mapping map. Uid is "not
relevant" and only present for uniqueness. Same args as
map_button_action otherwise.
- axis_mapping_item -> item of the axis mapping map. Uid is "not
relevant" and only present for uniqueness. Same args as
map_axis_action otherwise.
- application_axis_event -> If an axis is mapped to
application_action_X, this event will be called with the action as the
first argument, and the axis value as the second argument when the
axis is moved.
- application_button_event -> If a button is mapped to
application_action_X, this event will be called with the action as its
sole argument whent the action is triggered.

Change-Id: I6c551a76b4bc4593e8a2754e40b2fd7530ab07ee
Reviewed-on: https://gerrit.parrot.biz/35245
Reviewed-by: Jean-Baptiste Dubois <jean-baptiste.dubois@parrot.com>