[DEV] Added motion detection messages.
Motion detection messages will be used for thrown take off.

The app enables or disable the motion detection.
If enabled, motion state will be reported by the drone (at a
filtered rate).
If disabled, motion state is reported once, with state steady.

To use or not thrown takeoff, the app should enable motion
detection.
- If motion is steady, it should send a takeOff.
- If motion is moving, it should send a userTakeOff. In this
case, the drone will ramp its motors and send the flying state
motorRamping. When motors have reached the desired speed,
flying state will be set to UserTakeOff. Then, the drone will
wait to be thrown.

If the drone is steady and it receives a userTakeOff, it should
do the userTakeOff anyway.

Change-Id: I2c6ddd3bda514b19ef904d5b6417803a432f9aca
Reviewed-on: https://gerrit.parrot.biz/74043
Reviewed-by: Nicolas Christe <nicolas.christe@parrot.com>
Reviewed-by: Edouard Leurent <edouard.leurent@parrot.com>